requirements.txt
aiohttp==3.8.1
discord.py==2.0.0
python-dotenv==0.19.2
websockets==10.4